backend/pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Backend for GoAround project, CS300"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "alembic>=1.17.2",
    "fastapi[standard]>=0.121.1",
    "psycopg[binary]>=3.2.12",
    "pydantic>=2.12.4",
    "pydantic-settings>=2.12.0",
    "pyjwt>=2.10.1",
    "sqlalchemy>=2.0.44",
]

[dependency-groups]
dev = [
    "coverage>=7.11.3",
    "pytest>=9.0.0",
    "pytest-asyncio>=1.3.0",
    "pytest-postgresql>=7.0.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.coverage.run]
source = ["app"]
dynamic_context = "test_function"

[tool.coverage.report]
show_missing = true
sort = "-Cover"

[tool.coverage.html]
show_contexts = true

[tool.pytest.ini_options]
addopts = "-v"
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"
